fix: extract deep_research_workflow to own module to reduce definitions.py size
Move deep_research_workflow() from definitions.py into
factory/workflow/deep_research.py to reduce god file line count delta.
Prompt strings are now module-level constants instead of function-local,
reducing function complexity. Uses lazy import in the registry, matching
the pattern of deep_qa.py and research.py.
